backend/requirements.txt
fastapi>=0.104.1
fastapi-users[sqlalchemy]>=13.0.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
databases>=0.8.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
pillow>=10.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
slowapi>=0.1.9
email-validator>=2.0.0
python-dotenv>=1.0.0
bleach>=6.0.0